LayoutTests/fast/forms/fieldset/fieldset-type.html

<script src="../../js/resources/js-test-pre.js"></script>
<fieldset id="fs1" name="a"></fieldset>
<script>
description("This test checks fieldset.test attribute");
shouldBe('fs1.type', '"fieldset"');
</script>
<script src="../../js/resources/js-test-post.js"></script>
