third_party/blink/web_tests/fast/forms/fieldset/fieldset-type-expected.txt

This test checks fieldset.test attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS fs1.type is "fieldset"
PASS successfullyParsed is true

TEST COMPLETE
